(#56) Tidy Analyzer project
