+++ src/Stravaig.Extensions.Core.Analyzer.Tests/Sec0001/Sec0001StringIsNullOrWhiteSpaceAnalyzerUnitTest_Matches.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 using NUnit.Framework;
-using static Stravaig.Extensions.Core.Analyzer.Tests.CSharpAnalyzerVerifier<Stravaig.Extensions.Core.Analyzer.SEC0001_UseStringHasContentAnalyzer>;
+using static Stravaig.Extensions.Core.Analyzer.Tests.CSharpAnalyzerVerifier<Stravaig.Extensions.Core.Analyzer.Sec0001UseStringHasContentAnalyzer>;
 
 namespace Stravaig.Extensions.Core.Analyzer.Tests.Sec0001;
