+++ src/Stravaig.Extensions.Core.Analyzer.Tests/Sec0001/Sec0001StringIsNullOrWhiteSpaceAnalyzerUnitTest_NotMatches.cs
@@ -2,7 +2,7 @@
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Testing;
 using NUnit.Framework;
-using static Stravaig.Extensions.Core.Analyzer.Tests.CSharpAnalyzerVerifier<Stravaig.Extensions.Core.Analyzer.SEC0001_UseStringHasContentAnalyzer>;
+using static Stravaig.Extensions.Core.Analyzer.Tests.CSharpAnalyzerVerifier<Stravaig.Extensions.Core.Analyzer.Sec0001UseStringHasContentAnalyzer>;
 
 namespace Stravaig.Extensions.Core.Analyzer.Tests.Sec0001;
