+++ src/Stravaig.Extensions.Core.Analyzer.Tests/Sec0001/Sec0001StringIsNullOrWhiteSpaceCodeFixUnitTest_ExpressionFixes.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using NUnit.Framework;
 using static Stravaig.Extensions.Core.Analyzer.Tests.CSharpCodeFixVerifier<
-    Stravaig.Extensions.Core.Analyzer.SEC0001_UseStringHasContentAnalyzer,
+    Stravaig.Extensions.Core.Analyzer.Sec0001UseStringHasContentAnalyzer,
     Stravaig.Extensions.Core.Analyzer.SEC0001_UseStringHasContentAnalyzerCodeFixProvider>;
 
 namespace Stravaig.Extensions.Core.Analyzer.Tests.Sec0001;
