+++ src/Stravaig.Extensions.Core.Analyzer.Tests/Sec0011/Sec0011StringCompareReplaceWithIsBeforeCodeFixUnitTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using NUnit.Framework;
 using static Stravaig.Extensions.Core.Analyzer.Tests.CSharpCodeFixVerifier<
-    Stravaig.Extensions.Core.Analyzer.SEC001x_ReplaceStringCompareAnalyzer,
+    Stravaig.Extensions.Core.Analyzer.Sec001XReplaceStringCompareAnalyzer,
     Stravaig.Extensions.Core.Analyzer.SEC0011_ReplaceStringCompareWithBeforeAnalyzerCodeFix>;
 
 namespace Stravaig.Extensions.Core.Analyzer.Tests.Sec0011;
