+++ src/Stravaig.Extensions.Core.Analyzer.Tests/Sec0012/Sec0012ReplaceStringCompareAnalyzerTests_Matches.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 using NUnit.Framework;
-using static Stravaig.Extensions.Core.Analyzer.Tests.CSharpAnalyzerVerifier<Stravaig.Extensions.Core.Analyzer.SEC001x_ReplaceStringCompareAnalyzer>;
+using static Stravaig.Extensions.Core.Analyzer.Tests.CSharpAnalyzerVerifier<Stravaig.Extensions.Core.Analyzer.Sec001XReplaceStringCompareAnalyzer>;
 
 namespace Stravaig.Extensions.Core.Analyzer.Tests.Sec0012;
