+++ src/Stravaig.Extensions.Core.Analyzer.Tests/Sec0012/Sec0012StringCompareReplaceWithIsBeforeOrEqualToCodeFixUnitTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using NUnit.Framework;
 using static Stravaig.Extensions.Core.Analyzer.Tests.CSharpCodeFixVerifier<
-    Stravaig.Extensions.Core.Analyzer.SEC001x_ReplaceStringCompareAnalyzer,
+    Stravaig.Extensions.Core.Analyzer.Sec001XReplaceStringCompareAnalyzer,
     Stravaig.Extensions.Core.Analyzer.SEC0012_ReplaceStringCompareWithBeforeOrEqualAnalyzerCodeFix>;
 
 namespace Stravaig.Extensions.Core.Analyzer.Tests.Sec0012;
