+++ src/Stravaig.Extensions.Core.Analyzer/AnalyzerReleases.Shipped.md
@@ -3,6 +3,7 @@
 
 ## Release 0.1.1
 
+
 | Rule ID | Category    | Severity | Notes                                                  |
 |---------|-------------|----------|--------------------------------------------------------|
 | SEC0001 | Readability | Warning  | SEC0001_UseStringHasContentAnalyzer; [Documentation]() |
