+++ src/Stravaig.Extensions.Core.Analyzer/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Stravaig.Extensions.Core.Analyzer.Tests")]
