+++ src/Stravaig.Extensions.Core.Analyzer/Category.cs
@@ -1,0 +1,6 @@
+namespace Stravaig.Extensions.Core.Analyzer;
+
+internal static class Category
+{
+    internal const string Readability = "Readability";
+}
